requirements.txt
# Для визуализации (опционально)
matplotlib>=3.5.0

# FastAPI и зависимости
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Для работы с аудио (уже установлены в core)
# numpy>=1.21.0
# scipy>=1.7.0
# librosa>=0.9.0
# soundfile>=0.11.0
# noisereduce>=1.0.0

# Для дополнительных функций
python-magic>=0.4.27  # Для определения MIME типа
aiofiles>=23.0.0  # Для асинхронной работы с файлами

# Telegram бот
python-telegram-bot[job-queue]==20.7

# Асинхронные HTTP запросы
aiohttp>=3.9.0

uvicorn[standard]==0.24.0
python-multipart==0.0.6